+++ help.sh
@@ -9,7 +9,7 @@
 
 fg_current_year=$(date +%Y)
 
-echo " Fancy Git v$FANCYGIT_VERSION (c) $fg_current_year by Diogo Alexsander Cavilha <diogocavilha@gmail.com>.
+echo " Fancy Git v$FANCYGIT_VERSION - $fg_current_year by Diogo Alexsander Cavilha <diogocavilha@gmail.com>.
  Prompt changer to show a few cool git informations about your repository on terminal.
 
  COMMANDS