+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### v5.2.3
 
 - Improve base path definitions
+- Improve code style
 
 ### v5.2.2